requirements.txt
py-cpuinfo
psutil
